+++ src/paths/floyd_warshall.c
@@ -20,6 +20,8 @@
 #include "igraph_interface.h"
 #include "igraph_stack.h"
 
+#include "core/interruption.h"
+#include "internal/utils.h"
 
 static igraph_error_t igraph_distances_floyd_warshall_original(
         const igraph_t *graph, igraph_matrix_t *res,
@@ -28,6 +30,8 @@ static igraph_error_t igraph_distances_floyd_warshall_original(
     igraph_integer_t no_of_nodes = igraph_vcount(graph);
 
     for (igraph_integer_t k = 0; k < no_of_nodes; k++) {
+        IGRAPH_ALLOW_INTERRUPTION();
+
         /* Iteration order matters for performance!
          * First j, then i, because matrices are stored as column-major. */
         for (igraph_integer_t j = 0; j < no_of_nodes; j++) {
@@ -58,19 +62,34 @@ static igraph_error_t igraph_distances_floyd_warshall_tree(
         const igraph_t *graph, igraph_matrix_t *res,
         const igraph_vector_t *weights) {
 
+    /* This is the "Tree" algorithm of Brodnik et al.
+     * A difference from the paper is that instead of using the OUT_k tree of shortest
+     * paths _starting_ in k, we use the IN_k tree of shortest paths _ending_ in k.
+     * This makes it easier to iterate through matrices in column-major order,
+     * i.e. storage order, thus increasing performance. */
+
     igraph_integer_t no_of_nodes = igraph_vcount(graph);
 
-    /* predecessors[u][v] is the second but last vertex on the shortest path from u to v */
-    igraph_matrix_int_t predecessors;
-    IGRAPH_MATRIX_INT_INIT_FINALLY(&predecessors, no_of_nodes, no_of_nodes);
+    /* successors[v][u] is the second vertex on the shortest path from v to u,
+       i.e. the parent of v in the IN_u tree. */
+    igraph_matrix_int_t successors;
+    IGRAPH_MATRIX_INT_INIT_FINALLY(&successors, no_of_nodes, no_of_nodes);
 
-    /* children[u][i] is the i-th child of u in a tree of shortest distances
-       rooted at k in the main loop below */
-    igraph_matrix_int_t children;
-    IGRAPH_MATRIX_INT_INIT_FINALLY(&children, no_of_nodes, no_of_nodes);
+    /* children[children_start[u] + i] is the i-th child of u in a tree of shortest paths
+       rooted at k, and ending in k, in the main loop below (IN_k). There are no_of_nodes-1
+       child vertices in total, as the root vertex is excluded. This is essentially a contiguously
+       stored adjacency list representation of IN_k. */
+    igraph_vector_int_t children;
+    IGRAPH_VECTOR_INT_INIT_FINALLY(&children, no_of_nodes-1);
 
-    /* no_of_children[u] is the number of children of u
-       rooted at k in the main loop below */
+    /* children_start[u] indicates where the children of u are stored in children[].
+       These are effectively the cumulative sums of no_of_children[], with the first
+       element being 0. The last element, children_start[no_of_nodes], is equal to the
+       total number of children in the tree, i.e. no_of_nodes-1. */
+    igraph_vector_int_t children_start;
+    IGRAPH_VECTOR_INT_INIT_FINALLY(&children_start, no_of_nodes+1);
+
+    /* no_of_children[u] is the number of children that u has in IN_k in the main loop below. */
     igraph_vector_int_t no_of_children;
     IGRAPH_VECTOR_INT_INIT_FINALLY(&no_of_children, no_of_nodes);
 
@@ -84,27 +103,46 @@ static igraph_error_t igraph_distances_floyd_warshall_tree(
     igraph_stack_int_t stack;
     IGRAPH_STACK_INT_INIT_FINALLY(&stack, no_of_nodes);
 
-    for (igraph_integer_t v = 0; v < no_of_nodes; v++) {
-        for (igraph_integer_t u = 0; u < no_of_nodes; u++) {
-            MATRIX(predecessors, u, v) = u;
+    for (igraph_integer_t u = 0; u < no_of_nodes; u++) {
+        for (igraph_integer_t v = 0; v < no_of_nodes; v++) {
+            MATRIX(successors, v, u) = u;
         }
     }
 
     for (igraph_integer_t k = 0; k < no_of_nodes; k++) {
-        /* resetting no_of_children vector */
-        igraph_vector_int_null(&no_of_children);
+        IGRAPH_ALLOW_INTERRUPTION();
 
-        /* constructing the tree out_k (as in the paper) but
-           representing it as the children matrix */
+        /* Count the children of each node in the shortest path tree, assuming that at
+           this point all elements of no_of_children[] are zeros. */
         for (igraph_integer_t v = 0; v < no_of_nodes; v++) {
-            if (k == v) {
-                continue;
-            }
-            igraph_integer_t parent = MATRIX(predecessors, k, v);
-            MATRIX(children, parent, VECTOR(no_of_children)[parent]) = v;
+            if (v == k) continue;
+            igraph_integer_t parent = MATRIX(successors, v, k);
             VECTOR(no_of_children)[parent]++;
         }
-        /* constructing dfs-traversal and dfs-skip arrays for the out_k tree */
+
+        /* Note: we do not use igraph_vector_int_cumsum() here as that function produces
+           an output vector of the same length as the input vector. Here we need an output
+           one longer, with a 0 being prepended to what vector_cumsum() would produce. */
+        igraph_integer_t cumsum = 0;
+        for (igraph_integer_t v = 0; v < no_of_nodes; v++) {
+            VECTOR(children_start)[v] = cumsum;
+            cumsum += VECTOR(no_of_children)[v];
+        }
+        VECTOR(children_start)[no_of_nodes] = cumsum;
+
+        /* Constructing the tree IN_k (as in the paper) and representing it
+           as a contiguously stored adjacency list. The entries of the no_of_children
+           vector as re-used as an index of where to insert child node indices.
+           At the end of the calculation, all elements of no_of_children[] will be zeros,
+           making this vector ready for the next iteration of the outer loop. */
+        for (igraph_integer_t v = 0; v < no_of_nodes; v++) {
+            if (v == k) continue;
+            igraph_integer_t parent = MATRIX(successors, v, k);
+            VECTOR(no_of_children)[parent]--;
+            VECTOR(children)[ VECTOR(children_start)[parent] + VECTOR(no_of_children)[parent] ] = v;
+        }
+
+        /* constructing dfs-traversal and dfs-skip arrays for the IN_k tree */
         IGRAPH_CHECK(igraph_stack_int_push(&stack, k));
         igraph_integer_t counter = 0;
         while (!igraph_stack_int_empty(&stack)) {
@@ -115,13 +153,14 @@ static igraph_error_t igraph_distances_floyd_warshall_tree(
                 /* a negative marker -parent - 1 that is popped right after
                    all the descendants of the parent were processed */
                 IGRAPH_CHECK(igraph_stack_int_push(&stack, -parent - 1));
-                for (igraph_integer_t l = 0; l < VECTOR(no_of_children)[parent]; l++) {
-                    IGRAPH_CHECK(igraph_stack_int_push(&stack, MATRIX(children, parent, l)));
+                for (igraph_integer_t l = VECTOR(children_start)[parent]; l < VECTOR(children_start)[parent + 1]; l++) {
+                    IGRAPH_CHECK(igraph_stack_int_push(&stack, VECTOR(children)[l]));
                 }
             } else {
                 VECTOR(dfs_skip)[-(parent + 1)] = counter;
             }
         }
+
         /* main inner loop */
         for (igraph_integer_t i = 0; i < no_of_nodes; i++) {
             igraph_real_t dki = MATRIX(*res, k, i);
@@ -135,7 +174,7 @@ static igraph_error_t igraph_distances_floyd_warshall_tree(
                 igraph_real_t dd = MATRIX(*res, j, i);
                 if (di < dd) {
                     MATRIX(*res, j, i) = di;
-                    MATRIX(predecessors, i, j) = MATRIX(predecessors, k, j);
+                    MATRIX(successors, j, i) = MATRIX(successors, j, k);
                     counter++;
                 } else {
                     counter = VECTOR(dfs_skip)[j];
@@ -148,13 +187,14 @@ static igraph_error_t igraph_distances_floyd_warshall_tree(
         }
     }
 
-    igraph_matrix_int_destroy(&predecessors);
-    igraph_matrix_int_destroy(&children);
     igraph_stack_int_destroy(&stack);
-    igraph_vector_int_destroy(&no_of_children);
-    igraph_vector_int_destroy(&dfs_skip);
     igraph_vector_int_destroy(&dfs_traversal);
-    IGRAPH_FINALLY_CLEAN(6);
+    igraph_vector_int_destroy(&dfs_skip);
+    igraph_vector_int_destroy(&no_of_children);
+    igraph_vector_int_destroy(&children_start);
+    igraph_vector_int_destroy(&children);
+    igraph_matrix_int_destroy(&successors);
+    IGRAPH_FINALLY_CLEAN(7);
 
     return IGRAPH_SUCCESS;
 }
@@ -177,6 +217,12 @@ static igraph_error_t igraph_distances_floyd_warshall_tree(
  * running times for most instances. See the reference below for more information.
  *
  * </para><para>
+ * Note that internally this function always computes the distance matrix
+ * for all pairs of vertices. The \p from and \p to parameters only serve
+ * to subset this matrix, but do not affect the time or memory taken by the
+ * calculation.
+ *
+ * </para><para>
  * Reference:
  *
  * </para><para>
@@ -187,6 +233,8 @@ static igraph_error_t igraph_distances_floyd_warshall_tree(
  *
  * \param graph The graph object.
  * \param res An intialized matrix, the distances will be stored here.
+ * \param from The source vertices.
+ * \param to The target vertices.
  * \param weights The edge weights. If \c NULL, all weights are assumed to be 1.
  *   Negative weights are allowed, but the graph must not contain negative cycles.
  * \param mode The type of shortest paths to be use for the
@@ -219,14 +267,15 @@ static igraph_error_t igraph_distances_floyd_warshall_tree(
  *
  * Time complexity:
  * The original variant has complexity O(|V|^3 + |E|).
- * Brodnik et al indicates expected-case complexity of O(|V|^2 log^2 |V|)
- * for the "Tree" variant.
+ * The "Tree" variant has expected-case complexity of O(|V|^2 log^2 |V|)
+ * according to Brodnik et al., while its worst-time complexity remains O(|V|^3).
  * Here |V| denotes the number of vertices and |E| is the number of edges.
  */
 igraph_error_t igraph_distances_floyd_warshall(
-    const igraph_t *graph, igraph_matrix_t *res,
-    const igraph_vector_t *weights, igraph_neimode_t mode,
-    const igraph_floyd_warshall_algorithm_t method) {
+        const igraph_t *graph, igraph_matrix_t *res,
+        igraph_vs_t from, igraph_vs_t to,
+        const igraph_vector_t *weights, igraph_neimode_t mode,
+        const igraph_floyd_warshall_algorithm_t method) {
 
     igraph_integer_t no_of_nodes = igraph_vcount(graph);
     igraph_integer_t no_of_edges = igraph_ecount(graph);
@@ -290,6 +339,12 @@ igraph_error_t igraph_distances_floyd_warshall(
         }
     }
 
+    /* If there are zero or one vertices, nothing needs to be done.
+     * This is special-cased so that at later stages we can rely on no_of_nodes - 1 >= 0. */
+    if (no_of_nodes <= 1) {
+        return IGRAPH_SUCCESS;
+    }
+
     switch (method) {
     case IGRAPH_FLOYD_WARSHALL_ORIGINAL:
         IGRAPH_CHECK(igraph_distances_floyd_warshall_original(graph, res, weights));
@@ -301,5 +356,8 @@ igraph_error_t igraph_distances_floyd_warshall(
     default:
         IGRAPH_ERROR("Invalid method.", IGRAPH_EINVAL);
     }
+
+    IGRAPH_CHECK(igraph_i_matrix_subset_vertices(res, graph, from, to));
+
     return IGRAPH_SUCCESS;
 }